fix(detector): PanelExpansion Int-Overflow bei Default-maxStep (Panels kollabierten auf 1px)

+++ src/main/kotlin/com/panela/comiccutter/PanelExpansion.kt
@@ -10,14 +10,15 @@ object PanelExpansion {
         return panels.mapIndexed { i, p ->
             val pl = p.x; val pt = p.y; val pr = p.x + p.width; val pb = p.y + p.height
             val others = panels.filterIndexed { j, _ -> j != i }
+            // maxStep-Cap in Long, sonst läuft pr + Int.MAX_VALUE (Default) über und kollabiert das Panel.
             val rNb = others.filter { overlap(pt, pb, it.y, it.y + it.height) && it.x >= pr }.minOfOrNull { it.x }
-            val right = (rNb?.let { pr + (it - pr) / 2 } ?: pageW).let { minOf(it, pr + maxStep) }
+            val right = (rNb?.let { pr + (it - pr) / 2 } ?: pageW).let { minOf(it.toLong(), pr.toLong() + maxStep).toInt() }
             val lNb = others.filter { overlap(pt, pb, it.y, it.y + it.height) && it.x + it.width <= pl }.maxOfOrNull { it.x + it.width }
-            val left = (lNb?.let { pl - (pl - it) / 2 } ?: 0).let { maxOf(it, pl - maxStep) }
+            val left = (lNb?.let { pl - (pl - it) / 2 } ?: 0).let { maxOf(it.toLong(), pl.toLong() - maxStep).toInt() }
             val bNb = others.filter { overlap(pl, pr, it.x, it.x + it.width) && it.y >= pb }.minOfOrNull { it.y }
-            val bottom = (bNb?.let { pb + (it - pb) / 2 } ?: pageH).let { minOf(it, pb + maxStep) }
+            val bottom = (bNb?.let { pb + (it - pb) / 2 } ?: pageH).let { minOf(it.toLong(), pb.toLong() + maxStep).toInt() }
             val tNb = others.filter { overlap(pl, pr, it.x, it.x + it.width) && it.y + it.height <= pt }.maxOfOrNull { it.y + it.height }
-            val top = (tNb?.let { pt - (pt - it) / 2 } ?: 0).let { maxOf(it, pt - maxStep) }
+            val top = (tNb?.let { pt - (pt - it) / 2 } ?: 0).let { maxOf(it.toLong(), pt.toLong() - maxStep).toInt() }
             PanelRect(left, top, (right - left).coerceAtLeast(1), (bottom - top).coerceAtLeast(1))
         }
     }
